imported Vertexset implementation

+++ src/bitmap_custom.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <string.h>
 
 #define ulong_bits 64
 #define ulong_mask &63
@@ -17,7 +18,7 @@ static inline bool Bitmap_Test(unsigned long long *array, uint32_t vertex){
 }
 
 static inline void Bitmap_Clean(unsigned long long *array, size_t size){
-    memset(array,0,size*sizeof(unsigned long));
+    memset(array, 0, size * sizeof(unsigned long long));
 }
